Split DevicesController: extract archive/delete helpers
